third_party/WebKit/LayoutTests/fast/dom/HTMLAnchorElement/remove-href-from-focused-anchor.html

<!DOCTYPE html>
<body>
<a href="javascript:" id="target">link</a>
<script src="../../../resources/js-test.js"></script>
<script>
jsTestIsAsync = true;
var target = document.getElementById('target');
target.focus();
shouldBe('document.activeElement', 'target');
debug('Remove href.');
window.onload = function() {
    target.removeAttribute('href');
    shouldBe('document.activeElement', 'document.body');
    target.remove();
    finishJSTest();
};
</script>
</body>
