third_party/WebKit/LayoutTests/fast/dom/HTMLAnchorElement/remove-href-from-focused-anchor-expected.txt

PASS document.activeElement is target
Remove href.
PASS document.activeElement is document.body
PASS successfullyParsed is true

TEST COMPLETE
